✨ back + WS communication + front store working as expected for relationships

+++ back/src/relationship/relationship.service.ts
@@ -60,15 +60,14 @@ export class RelationshipService {
     peerId: string,
     relationType: RelationshipType,
   ) {
-    // @WebSocketServer() server: Server;
     if (userId == peerId) {
       throw new BadRequestException('user and peer cannot have same id');
     }
     const user: User = await this.userService.findEntity(userId);
     const peer: User = await this.userService.findEntity(peerId);
     let relation = new Array<Relationship>(2);
-    let userRelation = null;
-    let peerRelation = null;
+    let userRelation = RelationshipType.none;
+    let peerRelation = RelationshipType.none;
     try {
       relation = await this.findRelation(user, peer);
       userRelation = relation[0].type;
@@ -76,33 +75,17 @@ export class RelationshipService {
     } catch {
       console.log('No relationship exists');
     } finally {
-      if (relation[1] && relation[1].type == RelationshipType.blocked) {
-        throw new ForbiddenException('user is blocked');
+
+      if (this.updateForbidden(userRelation, peerRelation, relationType)) {
+        throw new ForbiddenException('Cannot update relation');
       }
       switch (relationType) {
         case RelationshipType.sent: {
-          if (userRelation == RelationshipType.friend) {
-            throw new BadRequestException('User and peer are already friends');
-          }
           userRelation = RelationshipType.sent;
           peerRelation = RelationshipType.received;
-          this.gatewayService.server.to(userId).emit('relationship-update', {
-            peerId: peerId,
-            relationType: userRelation,
-          });
-          this.gatewayService.server.to(peerId).emit('relationship-update', {
-            peerId: userId,
-            relationType: peerRelation,
-          });
           break;
         }
         case RelationshipType.friend: {
-          //Friend accept request can only be sent by a user that received an invite
-          if (userRelation != RelationshipType.received) {
-            throw new BadRequestException(
-              'Cannot be friends: no pending friend request',
-            );
-          }
           userRelation = RelationshipType.friend;
           peerRelation = RelationshipType.friend;
           break;
@@ -133,7 +116,53 @@ export class RelationshipService {
           peer: user,
         });
       }
+      if (userRelation === RelationshipType.blocked) {
+        this.gatewayService.server.to(userId).emit('relationship-update', {
+          peerId: peerId,
+          type: userRelation,
+        });
+      } else {
+        this.gatewayService.server.to(userId).emit('relationship-update', {
+          peerId: peerId,
+          type: userRelation,
+        });
+        this.gatewayService.server.to(peerId).emit('relationship-update', {
+          peerId: userId,
+          type: peerRelation,
+        });
+      }
     }
+  }
+
+  updateForbidden(
+    userState: RelationshipType,
+    peerState: RelationshipType,
+    newState: RelationshipType,
+  ) {
+    switch (newState) {
+      case RelationshipType.sent: {
+        if (
+          userState == RelationshipType.none &&
+          peerState == RelationshipType.none
+        ) {
+          return false;
+        }
+        break;
+      }
+      case RelationshipType.friend: {
+        if (
+          userState == RelationshipType.received &&
+          peerState == RelationshipType.sent
+        ) {
+          return false;
+        }
+        break;
+      }
+      case RelationshipType.blocked: {
+        return false;
+      }
+    }
+    return true;
   }
 
   async deleteRelation(userId: string, peerId: string): Promise<void> {
@@ -141,7 +170,15 @@ export class RelationshipService {
     const peer: User = await this.userService.findEntity(peerId);
     const relation = await this.findRelation(user, peer);
     console.log('Relation:', relation);
-    await this.relationshipRepository.delete(relation[0]);
-    await this.relationshipRepository.delete(relation[1]);
+    if (relation[1].type === RelationshipType.blocked) {
+      relation[0].type = RelationshipType.none;
+      await this.relationshipRepository.save(relation[0]);
+      this.gatewayService.server.to(userId).emit('relationship-delete', peerId);
+    } else {
+      await this.relationshipRepository.delete(relation[0]);
+      await this.relationshipRepository.delete(relation[1]);
+      this.gatewayService.server.to(userId).emit('relationship-delete', peerId);
+      this.gatewayService.server.to(peerId).emit('relationship-delete', userId);
+    }
   }
 }
